requirements.txt
Django==5.0
google-genai>=1.0.0
python-dotenv==1.0.0
gunicorn>=21.2.0
supabase==2.28.0